Rework the test by use viewport
